implement POST /times
including validation of the times in the backend
and having the date/time now in a proper format as suggested value.

There is the input type datetime-local in HTML5,
which gives a date picker (and not always a time picker).
However, we can't change the localization for this.
So I opted to provide a default value
to show the required format and make it easier to enter.