add Dockerfile
and use a different host name for timetracker
to use in docker via a profile

Note that host.docker.internal does not work out of the box on Linux.
This needs add-host when creating the container.

We build the jar file in one layer
and then run it in a separate smaller layer.